add auth for notes routes

+++ server/src/controllers/notes.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from 'express';
 import NoteModel from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../util/assertIsDefined';
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authUserId);
+    const notes = await NoteModel.find({ userId: authUserId }).exec();
     // throw Error('Bazinga!!!!');
     res.status(200).json(notes);
   } catch (error) {
@@ -15,8 +18,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefined(authUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid Note ID');
     }
@@ -24,6 +29,9 @@ export const getNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, 'Note not found');
     }
+
+    if (!note.userId.equals(authUserId))
+      throw createHttpError(401, "You can't accesss this note");
     res.status(200).json(note);
   } catch (error) {
     next(error);
@@ -43,11 +51,14 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authUserId = req.session.userId;
   try {
+    assertIsDefined(authUserId);
     if (!title) {
       throw createHttpError(400, 'Note must have title');
     }
     const newNote = await NoteModel.create({
+      userId: authUserId,
       title: title,
       text: text,
     });
@@ -74,7 +85,9 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authUserId = req.session.userId;
   try {
+    assertIsDefined(authUserId);
     if (!mongoose.isValidObjectId(noteId))
       throw createHttpError(400, 'Invalid note id');
 
@@ -82,6 +95,8 @@ export const updateNote: RequestHandler<
 
     const note = await NoteModel.findById(noteId).exec();
     if (!note) throw createHttpError(404, 'note not found');
+    if (!note.userId.equals(authUserId))
+      throw createHttpError(401, "You can't accesss this note");
 
     note.title = newTitle;
     note.text = newText;
@@ -96,13 +111,17 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefined(authUserId);
     if (!mongoose.isValidObjectId(noteId))
       throw createHttpError(400, 'invalid note id');
 
     const note = await NoteModel.findById(noteId).exec();
     if (!note) throw createHttpError(404, 'Note not found');
+    if (!note.userId.equals(authUserId))
+      throw createHttpError(401, "You can't accesss this note");
 
     await note.deleteOne();
     res.sendStatus(204);
